Seperated EditLog component into EditLog, EditMap and EditPhoto components for readability.
