+++ public/index.html
@@ -12,7 +12,7 @@
 <link href="https://fonts.googleapis.com/css2?family=Kalam:wght@700&family=Mali:wght@300&display=swap" rel="stylesheet">
 
     
-    <title>App</title>
+    <title>ShroomLogger</title>
   </head>
   <body>
     <noscript>
